Добавил домашнюю работу по уроку 7
