+++ 07_lesson/pages/shop_page.py
